refactor(worker): extract agent-turn sandbox runtime and establish
Move lease heartbeat, mutation fencing, recording, and late-sandbox
release into createSandboxTurnRuntime, and route/establish/lazy
provisioning into resolveSandboxRoute, establishTurnSandbox, and
bindLazySandboxProvisioner.
